+++ src/pages/index.tsx
@@ -1,6 +1,6 @@
 import Home from '@/pages/Home';
 import SignUp from '@/pages/SignUp';
 import Search from '@/pages/Search';
-import Save from '@/pages/Save';
+import Save from '@/pages/save/Save';
 
 export { Home, SignUp, Search, Save };
